// src/interfaces/player.interface.ts
export interface Player {
    id: number;
    name: string;
    score: number;
    first_position: number;
    second_position: number;
    third_position: number;
    games: number;
    avg: number;
    mpr: number;
    closed: number;
    hat_trick: number;
    nine_nine: number;
    white_horse: number;
    perfect: number;
    avatar: string;
}